+++ skrub/__init__.py
@@ -15,6 +15,7 @@ from ._minhash_encoder import MinHashEncoder
 from ._multi_agg_joiner import MultiAggJoiner
 from ._reporting import TableReport
 from ._select_cols import DropCols, SelectCols
+from ._sentence_encoder import SentenceEncoder
 from ._similarity_encoder import SimilarityEncoder
 from ._table_vectorizer import TableVectorizer
 from ._tabular_learner import tabular_learner
@@ -48,4 +49,5 @@ __all__ = [
     "AggTarget",
     "SelectCols",
     "DropCols",
+    "SentenceEncoder",
 ]